requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-admin-interface==0.28.5
django-ckeditor==6.7.0
django-colorfield==0.11.0
django-crispy-forms==2.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-slugify==8.0.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.4
